+++ CHANGELOG.md
@@ -142,6 +142,7 @@
 - Suppression de l'URL dans l'onglet de retour du SAE
 - Suppression du support MySQL. La seule base supportée est MariaDB.
 - Suppression de la création d'une entité par API via PATCH avec le paramètre 'create' et suppression du test 'testCreateWithEditAction()'
+- Suppression de la classe `UtilisateurCreator`, remplacée par `UserCreationService`
 
 ## Limitations